add FrameManager and optimize viewport projection updates (#471)
This commit introduces a centralized FrameManager to manage the OpenGL
render loop. It provides start-to-start interval pacing to ensure 
consistent frame rates and provides an RAII 'Frame' object to signal 
paint completion.

Projection logic has been moved from MapCanvasData to ProjectionUtils. 
MapCanvasViewport now uses a 'dirty' flag to cache the ViewProj matrix, 
recalculating it only when scroll, scale, layer, or config changes.

Key changes:
- New: FrameManager handles QTimer-based heartbeats for animations.
- New: ProjectionUtils centralizes 3D perspective and old-style frustum math.
- Config: Added maximumFps setting to advanced canvas options.
- Cleanup: Standardized layer access via getCurrentLayer() and fixed 
  integer-to-float conversion warnings in rendering paths.

+++ src/display/Connections.cpp
@@ -221,7 +221,7 @@ void ConnectionDrawer::drawRoomDoorName(const RoomHandle &sourceRoom,
     });
 
     static const auto bg = Colors::black.withAlpha(0.4f);
-    const glm::vec3 pos{xy, m_currentLayer};
+    const glm::vec3 pos{xy, static_cast<float>(m_currentLayer)};
     m_roomNameBatch.emplace_back(GLText{pos,
                                         mmqt::toStdStringLatin1(name), // GL font is latin1
                                         Colors::white,
